refactor: move Queue bucketing into the shared package
The Queue's ordering rules lived in the desktop renderer, but the mobile
app needs the same Queue. Copying them would fork the definition of what
needs the user first across two clients, so `bucketQueueThreads` moves to
@polycode/shared alongside the snooze predicates it already depends on.

The module was always platform-free — its only imports were `isWoken` and
two types — so this is a move, not a rewrite.

packages/shared has no test runner of its own, so its suite is registered
as a vitest project in the desktop config. `pnpm test` therefore keeps
covering the ordering rules for both clients.
