+++ apps/desktop/src/renderer/src/components/sidebar/QueueSidebar.tsx
@@ -2,7 +2,7 @@ import { type ReactNode, useCallback, useEffect, useState } from 'react'
 import { Archive, ArchiveRestore, BellRing, ChevronDown, ChevronRight, Clock, Inbox, PanelLeft, Plus, Settings } from 'lucide-react'
 import { SidebarViewMode } from '../../stores/ui'
 import { QueueThread, ThreadStatus } from '../../types/ipc'
-import { bucketQueueThreads } from '../../lib/queueBuckets'
+import { bucketQueueThreads } from '@polycode/shared'
 import { getThreadStatusColor, relativeTime, SidebarResizeHandle, ViewModeSwitch } from './shared'
 import SnoozeMenu, { timeUntil } from './SnoozeMenu'
 import ProjectFavicon from '../ProjectFavicon'
