+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './types'
 export * from './snooze'
+export * from './queue-buckets'
 export * from './messages'
 export * from './todos'
 export * from './tools'
